requirements.txt
# Social Sentiment Pipeline Dependencies
# =======================================

# API Service - Flask
flask>=3.0.0
gunicorn>=21.0.0

# API Service - FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Database
psycopg2-binary>=2.9.9

# HTTP Client
aiohttp>=3.9.0
requests>=2.31.0

# Environment
python-dotenv>=1.0.0

# Optional: LLM integration for fallback classification
# Uncomment if using Gemini API for LLM-based sentiment classification
# google-generativeai>=0.3.0
